Basic Functionality Added
I have added some basic AJAX functionality, so now it only loads the
maps on the tabs that you actually click, rather than them all at once.
However, if you click on the same tab more than once, it will re-load it
again and again each time. This needs to be addressed in the future.